+++ lib/backup-run.js
@@ -1,0 +1,6 @@
+'use strict';
+
+require('dotenv').config({});
+var backup = require('./backup');
+
+backup();
