+++ lib/format-date.js
@@ -6,7 +6,7 @@ function formatDate(date) {
   if (!date) {
     date = new Date();
   }
-  return _formatDate(date, 'YYYY-MM-DD');
+  return _formatDate(date, 'YYYYMMDD-hhmmss');
 }
 
 module.exports = formatDate;
