+++ package.json
@@ -4,7 +4,7 @@
   "description": "Dump a MySQL database and sync the backup to AWS S3.",
   "main": "index.js",
   "scripts": {
-    "backup": "node index.js"
+    "backup": "node lib/backup-run.js"
   },
   "repository": {
     "type": "git",
@@ -17,6 +17,7 @@
     "@ryanburnette/parse-mysql-url": "^1.1.0",
     "aws-sdk": "^2.517.0",
     "date-fns": "^1.30.1",
+    "dotenv": "^8.2.0",
     "node-cron": "^2.0.3"
   }
 }
